# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "COZE-DISCORD-PROXY 代理服务",
        "title": "COZE-DISCORD-PROXY",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/api/channel/create": {
            "post": {
                "description": "创建频道",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channel"
                ],
                "summary": "创建频道",
                "parameters": [
                    {
                        "description": "channelModel",
                        "name": "channelModel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ChannelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/model.ChannelResp"
                        }
                    }
                }
            }
        },
        "/api/channel/del/{id}": {
            "delete": {
                "description": "删除频道",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channel"
                ],
                "summary": "删除频道",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.ChannelReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "type": {
                    "type": "number"
                }
            }
        },
        "model.ChannelResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}